+++ src/code/matrix/evaluationfunctions/MatrixGreedy2.java
@@ -6,7 +6,7 @@ import code.searchproblem.strategies.EvaluationFunction;
 public class MatrixGreedy2 extends EvaluationFunction {
 
     @Override
-    public int evaluate(SearchTreeNode node) {
+    public long evaluate(SearchTreeNode node) {
         return MatrixHeuristics.heuristic2(node);
     }
